refactor(session): extract useDesignCatalog from DesignSession
The examples / capabilities / trust cluster (examples, examplesError,
havePynec, terrainPresets, loadErrors, trustBusy, loadExamples, trustDesign,
the two mount fetches and the auto-select-geometry effect) moves whole into
components/session/useDesignCatalog.ts. Bodies and every literal dep array
are unchanged; the three effects keep their relative order.

The only reflow is a comment: the "schema-driven parameter controls"
paragraph follows the /examples fetch into the hook, and the multi-band
paragraph now sits directly above paramValues, which stays behind.
